+++ src/routers/xcm/Router.sol
@@ -21,7 +21,7 @@ interface ConnectorLike {
     ) external;
     function updateMember(uint64 poolId, bytes16 trancheId, address user, uint64 validUntil) external;
     function updateTokenPrice(uint64 poolId, bytes16 trancheId, uint128 price) external;
-    function handleTransfer(uint64 poolId, bytes16 trancheId, address user, uint256 amount) external;
+    function handleTransfer(uint64 poolId, bytes16 trancheId, address user, uint128 amount) external;
 }
 
 contract ConnectorXCMRouter {
@@ -68,7 +68,7 @@ contract ConnectorXCMRouter {
             (uint64 poolId, bytes16 trancheId, uint128 price) = ConnectorMessages.parseUpdateTokenPrice(_msg);
             connector.updateTokenPrice(poolId, trancheId, price);
         } else if (ConnectorMessages.isTransfer(_msg)) {
-            (uint64 poolId, bytes16 trancheId,, address user, uint256 amount) = ConnectorMessages.parseTransfer(_msg);
+            (uint64 poolId, bytes16 trancheId,, address user, uint128 amount) = ConnectorMessages.parseTransfer(_msg);
             connector.handleTransfer(poolId, trancheId, user, amount);
         } else {
             require(false, "invalid-message");
